+++ apps/frontend/src/components/page-products/WatchedProducts.vue
@@ -1,8 +1,8 @@
 <script>
-import TileProduct from './TileProduct.vue'
+import ListPave from './lists/ListPave.vue'
 
 export default {
-  components: { TileProduct },
+  components: { ListPave },
 
   props: ['products'],
 }
@@ -11,12 +11,7 @@ export default {
 <template>
   <div class="watched-products">
     <h3>Просмотренные товары</h3>
-    <ul class="products">
-      <TileProduct
-        v-for="product of products"
-        :key="product.id"
-        :product="product"
-      />
-    </ul>
+
+    <ListPave :products="products" />
   </div>
 </template>
