+++ apps/frontend/src/components/page-products/lists/items/ItemBrick.vue
