+++ apps/frontend/src/components/page-products/lists/items/ItemTile.vue
