+++ apps/frontend/src/components/page-products/lists/items/ItemTr.vue
